Move StatusBox component to separate file
